Split functions into model, view, controller folders that follow the MVC pattern. All requests pass through index.php which provides the user with views and updates the database using model functions.
